Remove unused `client/` routes
These were added as part of the Oauth implementation,
but they are not used.

For the most part, all of this code was taken from
https://github.com/authlib/playground/;
it allows users to register client apps through the UI.

Since this code was added, we've settled on registering client app
credentials directly into the database.

Whether or not that approach is the best might be up for debate,
but at the moment this code is unused and untested.
If we decide that a client registration UI would be nice in the future,
we can always revert this commit.

See the current UI rendered at
<https://tb-api-dev.cirg.washington.edu/client>.